Extract data table into reusable component
